Moved all tests for models
