+++ src/models/AmbulanceTest.java
@@ -1,4 +1,4 @@
-package models.tests;
+package models;
 
 import models.Ambulance;
 import models.Point;
