+++ src/models/PatientTest.java
@@ -1,4 +1,4 @@
-package models.tests;
+package models;
 
 import models.InjurySeverity;
 import models.Patient;
